+++ gnomad/assessment/sanity_checks.py
@@ -48,13 +48,13 @@ def generic_field_check(
     ht = ht.filter(cond_expr)
     n_fail = ht.count()
     if n_fail > 0:
-        logger.info(f"Found {n_fail} sites that fail {check_description} check:")
+        logger.info("Found %d sites that fail %s check:", n_fail, check_description)
         if show_percent_sites:
-            logger.info(f"Percentage of sites that fail: {n_fail / ht_orig.count()}")
+            logger.info("Percentage of sites that fail: %f", n_fail / ht_orig.count())
         ht = ht.flatten()
         ht.select("locus", "alleles", *display_fields).show()
     else:
-        logger.info(f"PASSED {check_description} check")
+        logger.info("PASSED %s check", check_description)
         if verbose:
             ht_orig = ht_orig.flatten()
             ht_orig.select(*display_fields).show()
@@ -188,5 +188,5 @@ def compare_row_counts(ht1: hl.Table, ht2: hl.Table) -> bool:
     """
     r_count1 = ht1.count()
     r_count2 = ht2.count()
-    logger.info(f"{r_count1} rows in left table; {r_count2} rows in right table")
+    logger.info("%d rows in left table; %d rows in right table", r_count1, r_count2)
     return r_count1 == r_count2