+++ gnomad/resources/grch38/gnomad.py
@@ -1,9 +1,9 @@
 # noqa: D100
 
 from gnomad.resources.resource_utils import (
-    TableResource,
+    GnomadPublicTableResource,
+    GnomadPublicMatrixTableResource,
     VersionedMatrixTableResource,
-    MatrixTableResource,
     VersionedTableResource,
     DataException,
 )
@@ -31,7 +31,7 @@ GROUPS = ["adj", "raw"]
 SEXES = ["XX", "XY"]
 POPS = ["afr", "ami", "amr", "asj", "eas", "fin", "nfe", "oth", "sas", "mid"]
 COHORTS_WITH_POP_STORED_AS_SUBPOP = ["tgp", "hgdp"]
-KG_POPS = [
+TGP_POPS = [
     "esn",
     "pur",
     "pjl",
@@ -113,6 +113,7 @@ HGDP_POPS = [
     "burusho",
     "maya",
 ]
+POPS_STORED_AS_SUBPOPS = TGP_POPS + HGDP_POPS
 POPS_TO_REMOVE_FOR_POPMAX = {"asj", "fin", "oth", "ami", "mid"}
 DOWNSAMPLINGS = [
     10,
@@ -146,7 +147,7 @@ DOWNSAMPLINGS = [
 gnomad_syndip = VersionedMatrixTableResource(
     default_version="3.0",
     versions={
-        "3.0": MatrixTableResource(
+        "3.0": GnomadPublicMatrixTableResource(
             path="gs://gnomad-public/truth-sets/hail-0.2/gnomad_v3_syndip.b38.mt"
         )
     },
@@ -155,7 +156,7 @@ gnomad_syndip = VersionedMatrixTableResource(
 na12878 = VersionedMatrixTableResource(
     default_version="3.0",
     versions={
-        "3.0": MatrixTableResource(
+        "3.0": GnomadPublicMatrixTableResource(
             path="gs://gnomad-public/truth-sets/hail-0.2/gnomad_v3_na12878.mt"
         )
     },
@@ -208,7 +209,9 @@ def public_release(data_type: str) -> VersionedTableResource:
     return VersionedTableResource(
         current_release,
         {
-            release: TableResource(path=_public_release_ht_path(data_type, release))
+            release: GnomadPublicTableResource(
+                path=_public_release_ht_path(data_type, release)
+            )
             for release in releases
         },
     )
@@ -236,7 +239,9 @@ def coverage(data_type: str) -> VersionedTableResource:
     return VersionedTableResource(
         current_release,
         {
-            release: TableResource(path=_public_coverage_ht_path(data_type, release))
+            release: GnomadPublicTableResource(
+                path=_public_coverage_ht_path(data_type, release)
+            )
             for release in releases
         },
     )